added WeatherstackCurrentWeatherClient

+++ src/main/java/pl/sda/magcie/weather/httpclient/accuweather/MetricDTO.java
@@ -1,5 +1,5 @@
 package pl.sda.magcie.weather.httpclient.accuweather;
 
 class MetricDTO {
-    public int Value;
+    public double Value;
 }
